Move tests to correct namespace
Move the test classes to reflect the change in the original class
