+++ tests/Mutator/Util/MutatorProfileTest.php
@@ -7,7 +7,7 @@
 
 declare(strict_types=1);
 
-namespace Infection\Tests\Mutator;
+namespace Infection\Tests\Mutator\Util;
 
 use Infection\Mutator\Util\MutatorProfile;
 use PHPUnit\Framework\TestCase;
